refactor: Вынести функцию в functions.php
